CDD-2567: Renamed the risk matrix file and minimised the image
